+++ component/component.go
@@ -32,7 +32,7 @@ func NewSession(ctx context.Context, addr jid.JID, secret []byte, rw io.ReadWrit
 	if received {
 		return xmpp.ReceiveSession(ctx, rw, Negotiator(addr, secret, received))
 	}
-	return xmpp.NegotiateSession(ctx, addr, addr, rw, Negotiator(addr, secret, received))
+	return xmpp.NewSession(ctx, addr, addr, rw, Negotiator(addr, secret, received))
 }
 
 // Negotiator returns a new function that can be used to negotiate a component